requirements.txt
django
djangorestframework
python-dotenv
django-cors-headers
pymongo
gunicorn
langchain
langchain-community
langchain-openai
chromadb==0.4.22
openai